requirements.txt
numpy
scipy
scikit-learn
joblib
pyusb
pyzmq
matplotlib
pyrtlsdr
pyshark
tk
libusb
asyncio